Move ModuleScopeDimKeywordInspectionTests for QuickFixes into relevant files under RD.QuickFixes
