+++ RubberduckTests/QuickFixes/IgnoreOnceQuickFixTests.cs
@@ -412,6 +412,29 @@ End Sub";
             Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
 
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void ModuleScopeDimKeyword_IgnoreQuickFixWorks()
+        {
+            const string inputCode =
+@"Dim foo";
+
+            const string expectedCode =
+@"'@Ignore ModuleScopeDimKeyword
+Dim foo";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new ModuleScopeDimKeywordInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            new IgnoreOnceQuickFix(state, new[] { inspection }).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
 
     }
 }
